# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODO List API
  description: |
    A comprehensive RESTful API for managing tasks organized within multiple lists.

    ## Features
    - Create and manage multiple todo lists
    - Add, update, and delete tasks within lists
    - Set deadlines and priorities for tasks
    - Track completion status
    - Advanced querying and filtering
    - Task statistics and productivity insights

    ## Authentication
    Currently, this API does not require authentication. All endpoints are publicly accessible.

    ## Rate Limiting
    API requests are rate-limited to 1000 requests per hour per IP address.

    ## Data Validation
    All request bodies are validated against JSON schemas. Invalid requests will return 400 Bad Request with validation details.
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@todoapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.todolist.example.com
    description: Production server
security: []
...
paths:
...
  /api/lists:
    get:
      tags: [Lists]
      summary: Get all lists
      description: Retrieve all todo lists with optional filtering and sorting
      operationId: getLists
      parameters:
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [name, createdAt, updatedAt]
          description: Field to sort by
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
      responses:
        '200':
          description: Lists retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Lists retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Lists]
      summary: Create a new list
      description: Create a new todo list
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListRequest'
            example:
              name: 'Work Tasks'
              description: 'Tasks related to work projects'
      responses:
        '201':
          description: List created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'List created successfully'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/lists/{listId}:
    get:
      tags: [Lists]
      summary: Get a specific list
      description: Retrieve a list by its ID
      operationId: getListById
      parameters:
        - $ref: '#/components/parameters/ListIdParam'
      responses:
        '200':
          description: List retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'List retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Lists]
      summary: Update a list
      description: Update an existing list by its ID
      operationId: updateList
      parameters:
        - $ref: '#/components/parameters/ListIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListRequest'
            example:
              name: 'Updated Work Tasks'
              description: 'Updated description for work tasks'
      responses:
        '200':
          description: List updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'List updated successfully'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Lists]
      summary: Delete a list
      description: Delete a list and all its tasks
      operationId: deleteList
      parameters:
        - $ref: '#/components/parameters/ListIdParam'
      responses:
        '200':
          description: List deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                        example: true
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'List deleted successfully'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/lists/{listId}/tasks:
    get:
      tags: [Lists]
      summary: Get tasks in a list
      description: Retrieve all tasks within a specific list
      operationId: getTasksByList
      parameters:
        - $ref: '#/components/parameters/ListIdParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [completed, pending]
          description: Filter tasks by completion status
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filter tasks by priority
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [title, priority, deadline, createdAt, updatedAt]
          description: Field to sort by
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Tasks retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Tasks]
      summary: Create a task in a list
      description: Create a new task within a specific list
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/ListIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            example:
              title: 'Complete project documentation'
              description: 'Write comprehensive documentation for the new feature'
              priority: 'high'
              deadline: '2025-07-10T17:00:00.000Z'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TaskResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Task created successfully'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Task endpoints
...
components:
  parameters:
    ListIdParam:
      in: path
      name: listId
      required: true
      schema:
        type: string
        format: uuid
      description: The unique identifier of the list
      example: '123e4567-e89b-12d3-a456-426614174000'
...
  schemas:
...
    ListResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the list
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          description: Name of the list
          example: 'Work Tasks'
        description:
          type: string
          nullable: true
          description: Optional description of the list
          example: 'Tasks related to work projects'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the list was created
          example: '2025-07-01T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the list was last updated
          example: '2025-07-02T14:30:00.000Z'
        taskCount:
          type: integer
          description: Number of tasks in the list
          example: 5
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
          example: '987fcdeb-51a2-43d7-8f9e-123456789abc'
        listId:
          type: string
          format: uuid
          description: ID of the list this task belongs to
          example: '123e4567-e89b-12d3-a456-426614174000'
        title:
          type: string
          description: Title of the task
          example: 'Complete project documentation'
        description:
          type: string
          nullable: true
          description: Optional detailed description of the task
          example: 'Write comprehensive documentation for the new feature'
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Priority level of the task
          example: 'high'
        deadline:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task
          example: '2025-07-10T17:00:00.000Z'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: '2025-07-01T11:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: '2025-07-02T15:00:00.000Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the task was completed
          example: null

    # Request Models
    CreateListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the list
          example: 'Work Tasks'
        description:
          type: string
          maxLength: 500
          description: Optional description of the list
          example: 'Tasks related to work projects'
    UpdateListRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the list
          example: 'Updated Work Tasks'
        description:
          type: string
          maxLength: 500
          description: Optional description of the list
          example: 'Updated description for work tasks'
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Title of the task
          example: 'Complete project documentation'
        description:
          type: string
          maxLength: 1000
          description: Optional detailed description of the task
          example: 'Write comprehensive documentation for the new feature'
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
          description: Priority level of the task
          example: 'high'
        deadline:
          type: string
          format: date-time
          description: Optional deadline for the task
          example: '2025-07-10T17:00:00.000Z'
...
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: 'title'
                  message:
                    type: string
                    example: 'Title is required'
                  value:
                    oneOf:
                      - type: string
                      - type: number
                      - type: boolean
                      - type: 'null'
        timestamp:
          type: string
          format: date-time
    NotFoundError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'RESOURCE_NOT_FOUND'
            message:
              type: string
              example: 'The requested resource was not found'
            resource:
              type: string
              example: 'Task'
            id:
              type: string
              example: '123e4567-e89b-12d3-a456-426614174000'
        timestamp:
          type: string
          format: date-time
    InternalServerError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'INTERNAL_SERVER_ERROR'
            message:
              type: string
              example: 'An unexpected error occurred'
            requestId:
              type: string
              example: 'req_123456789'
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Bad Request - Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundError'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
...